requirements.txt
fastapi
uvicorn
gspread
google-auth
python-dotenv
requests
openai
